+++ components/CompareTable/Table.tsx
