+++ components/CompareTable/Title.tsx
