+++ components/CompareTable/index.tsx
