+++ pages/product.tsx
@@ -9,7 +9,7 @@ import {
   ProductDescription,
   IncludedProductsList,
   Reviews,
-  ProductCompareEditions
+  CompareTable
 } from 'components';
 import ProductSummary from 'components/ProductSummary';
 import { default as ProductInt } from 'lib/models/product';
@@ -90,7 +90,7 @@ class Products extends React.Component<{}, State> {
           </Container>
         </Box>
         <Container>
-          <ProductCompareEditions />
+          <CompareTable />
         </Container>
         <Box bg={containerBg2} pt={[3, 7]} pb={[4, 10]} mt={5}>
           <Container>
